Add changes: Create folders for representing 3-tier architecture: DataAccessLayer, BusinessLogicLayer, PresentaionLayer. Move activities and adapters into PresentationLayer, BusinesLogic file into BLL, DBHelper into DAL. Remove redundant comments and code in all layeres
