+++ app/src/main/res/layout/activity_history.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/login_bg"
-    tools:context=".activities.MainPageActivity">
+    tools:context=".PresentationLayer.Activities.MainPageActivity">
 
 
     <ImageView
