+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/login_bg"
-    tools:context=".activities.LoginActivity">
+    tools:context=".PresentationLayer.Activities.LoginActivity">
 
 
     <ImageView
