+++ app/src/main/res/layout/activity_sign_up.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/login_bg"
-    tools:context=".activities.LoginActivity">
+    tools:context=".PresentationLayer.Activities.LoginActivity">
 
     <ImageView
         android:id="@+id/trident_icon"
